requirements.txt
fastapi
uvicorn
python-dotenv
transformers==4.40.0
torch
sentence-transformers==2.6.1
langchain-community==0.3.26
langchain-core
langchain-huggingface
langchain-text-splitters
faiss-cpu
